+++ modules/hanna-react/src/utils.ts
@@ -1,4 +1,5 @@
 export * from './utils/useDidChange';
 export * from './utils/useFormatMonitor';
 export * from './utils/useGetSVGtext';
+export * from './utils/useMixedControlState';
 export * from './utils/useScrollbarWidthCSSVar';